+++ onlyoffice-docspace-wordpress-block/src/edit.js
@@ -55,6 +55,17 @@ const Edit = ({ attributes, setAttributes }) => {
         }
     ];
 
+    const editorTypes = [
+        {
+            label: __("Embedded", "onlyoffice-docspace-plugin"),
+            value: "embedded"
+        },
+        {
+            label: __("Editor", "onlyoffice-docspace-plugin"),
+            value: "desktop"
+        }
+    ];
+
     const script = () => {
         if (isOpen) {
             wp.oodsp.initLoginManager(
@@ -137,6 +148,7 @@ const Edit = ({ attributes, setAttributes }) => {
             height: "100%",
             mode: mode,
             selectorType: "roomsOnly",
+            theme: "Base",
             locale: _oodsp.locale,
             events: {
                 onSelectCallback: onSelectCallback,
@@ -197,6 +209,17 @@ const Edit = ({ attributes, setAttributes }) => {
                                 options={themes}
                                 onChange={(value) => {setAttributes({ theme: value })}}
                             />
+                            { 
+                                attributes.fileId ?
+                                    <SelectControl
+                                        label={__("View", "onlyoffice-docspace-plugin")}
+                                        value={attributes.editorType}
+                                        options={editorTypes}
+                                        onChange={(value) => {setAttributes({ editorType: value })}}
+                                    />
+                                    :
+                                    ''
+                            }
                         </PanelBody>
                     </InspectorControls>