+++ CHANGELOG.md
@@ -2,10 +2,22 @@
 
 
 ##
+## Added 
+- ability to add multiple rooms/files to a page
+- block settings (view mode 'editor/embedded')
+
+##
+## Changed
+- hide sign out button on page docspace
+- hide request name for anonymous
+- structure of tables with files (Name,Size,Type)
+- base theme in admin panel for docspace
+
+## 2.0.0
 ## Added
 - support for public rooms
-- settings for onlyoffice-docspace-wordpress-block (theme, align)
-- improoved view onlyoffice-docspace-wordpress-block in editor
+- improved block settings (theme, align)
+- improved view of the inserted blocks
 
 ## Changed
 - delete public user "Wordpress Viewer"